moved user from Entities\User to Entities\User\User
