+++ database/factories/UserFactory.php
@@ -13,7 +13,7 @@ use Faker\Generator as Faker;
 |
 */
 
-$factory->define(Hedonist\Entities\User::class, function (Faker $faker) {
+$factory->define(Hedonist\Entities\User\User::class, function (Faker $faker) {
     return [
         'name' => $faker->name,
         'email' => $faker->unique()->safeEmail,
