+++ tests/Unit/Auth/PresenterTest.php
@@ -5,7 +5,7 @@ namespace Tests\Unit\Auth;
 
 use Hedonist\Actions\Auth\Presenters\AuthPresenter;
 use Hedonist\Actions\Auth\Responses\AuthenticateResponseInterface;
-use Hedonist\Entities\User;
+use Hedonist\Entities\User\User;
 use Tests\TestCase;
 
 class PresenterTest extends TestCase
